requirements.txt
click==8.1.3
matplotlib==3.5.2
numpy==1.21.6
Pillow==9.1.1
python-dotenv==0.20.0
setuptools==41.2.0
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0